feature: Add some decoder and encoder tests.
Add some junit test for decoders and encoders below:
- int8
- uint8
- int16
- uint16
- uint32

+++ src/main/java/org/indunet/fastproto/annotation/type/UInteger32Type.java
@@ -17,7 +17,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface UInteger32Type {
     int SIZE = Integer.SIZE >> 3;
-    int MAX_VALUE = Integer.MAX_VALUE - Integer.MIN_VALUE;
+    long MAX_VALUE = (long) Integer.MAX_VALUE - Integer.MIN_VALUE;
     int MIN_VALUE = 0;
 
     int value();
